Changed all ID variables
Changed all ID variables' type to be size_t (they were int before that)

+++ include/Objects/Discounts.hpp
@@ -7,11 +7,11 @@
 class Discount
 {
 protected:
-    int discountID; // unique
+    size_t discountID; // unique
     float discountPercent;
 
 public:
-    Discount(int id, float percent);
+    Discount(size_t id, float percent);
     virtual ~Discount() = default;
 
     virtual float apply(const Fragrance &frag) const;
@@ -22,7 +22,7 @@ class BonusDiscount : public Discount
     float bonus;
 
 public:
-    BonusDiscount(int id, float percent, float bonus);
+    BonusDiscount(size_t id, float percent, float bonus);
 
     float apply(const Fragrance &frag) const override;
 };
@@ -32,6 +32,6 @@ class BrandDiscount : public Discount
     std::string brandName;
 
 public:
-    BrandDiscount(int id, float percent, const std::string &brandName);
+    BrandDiscount(size_t id, float percent, const std::string &brandName);
     float apply(const Fragrance &frag) const override;
 };
