+++ include/Objects/Fragrance.hpp
@@ -7,22 +7,22 @@
 
 class Fragrance
 {
-    int fragranceID; // unique
+    size_t fragranceID; // unique
     std::string name;
     std::string brand;
     float price;
-    std::vector<int> ingredientIDs;
+    std::vector<size_t> ingredientIDs;
     int quantity;
     std::vector<Review> reviews;
 
 public:
     Fragrance(const std::string &name, const std::string &brand, float price, const std::vector<int> &ingredientIDs);
 
-    int getID() const;
+    size_t getID() const;
     std::string getName() const;
     std::string getBrand() const;
     float getPrice() const;
-    const std::vector<int> &getIngredientIDs() const;
+    const std::vector<size_t> &getIngredientIDs() const;
     int getQuantity() const;
     const std::vector<Review> getReviews() const;
     float getAvgRating() const;
